Add cap button support to liblight, restore TW timeout behavior.
Adds capacitive button backlight support to liblight, which allows the
framework to turn on/off the backlight independent of screen on/off events.

Also disables automatic brightness for the button and keyboard backlights,
which, with the accompanying framework changes, restores the six-second
timeout behavior present in TW.

Auto brightness values for the button and keyboard are maxed so that, if
the framework changes are not included in a build, those lights stay on
regardless of light sensor values.

Change-Id: I366bc12bf72453c8592e0f561cf05325a33b4737